backend/ReviewService/requirements.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
